// src/app/model/IUserFields.tsx
export enum IUserFields {
  NAME = "Name",
  USERNAME = "Username",
  EMAIL = "Email",
  PHONE = "Phone",
  WEBSITE = "Website",
  CITY = "City",
  STREET = "Street",
  SUITE = "Suite",
  ZIPCODE = "Zip Code",
  COMPANYNAME = "Company Name",
  COMPANYCATCHPHRASE = "Company Cathphrase",
  COMPANYBS = "Company bs"
}